clean code again
