+++ srcs/execution/builtin/cwd/rm_backpath.c
@@ -6,7 +6,7 @@
 /*   By: rbroque <rbroque@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/24 17:31:32 by rbroque           #+#    #+#             */
-/*   Updated: 2023/04/24 18:04:19 by rbroque          ###   ########.fr       */
+/*   Updated: 2023/04/24 21:11:08 by rbroque          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,7 +74,7 @@ static void	cpy_clean_path(char *dest, char *src, size_t final_len)
 	dest[final_len] = '\0';
 }
 
-char	*get_clean_path(char *path)
+char	*rm_backpath(char *path)
 {
 	const size_t	len = get_clean_len(path);
 	char			*clean;
