+++ tests/CUNIT/srcs/minishell.test.c
@@ -42,7 +42,7 @@ void	set_tests(void)
 		{.suite = NULL, .sname = "Execution", .tests = {
 			{.tname = "cmd_path", cmd_path__test},
 			{.tname = "ft_getenv", ft_getenv__test},
-			{.tname = "change_var", change_var__test}
+			{.tname = "change_var", change_var__test},
 			}
 		},
 		{.suite = NULL, .sname = NULL}
